requirements.txt
args==0.1.0
certifi==2019.3.9
chardet==3.0.4
clint==0.5.1
coverage==4.5.3
cx-Oracle==7.1.3
dateparser==0.7.1
dateutils==0.6.6
dbms==1.1.1
idna==2.8
mysqlclient==1.4.2.post1
python-dateutil==2.8.0
pytz==2019.1
regex==2019.4.14
requests==2.21.0
six==1.12.0
tzlocal==1.5.1
urllib3==1.25.2
